Move and rename `use-server-effect`
