+++ packages/template-retail-react-app/app/pages/home/index.jsx
@@ -8,7 +8,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {useIntl, FormattedMessage} from 'react-intl'
-import UseServerEffect from 'pwa-kit-react-sdk/ssr/universal/hooks/use-server-effect'
+import UseServerEffect from 'pwa-kit-react-sdk/ssr/universal/server-effects'
 
 // TODO: Clean up how this default import is used.
 const {useServerEffect} = UseServerEffect
